+++ src/zope/schema/tests/test_fieldproperty.py
@@ -267,8 +267,9 @@ class FieldPropertyTests(_Base, _Integration):
 
     def test_field_event(self):
         from zope.schema import Text
-
+        from zope.interface.verify import verifyObject
         from zope.event import subscribers
+        from zope.schema.interfaces import IFieldUpdatedEvent
         from zope.schema.fieldproperty import FieldUpdatedEvent
         log = []
         subscribers.append(log.append)
@@ -283,18 +284,26 @@ class FieldPropertyTests(_Base, _Integration):
         self.assertEqual(len(log), 6)
         event = log[0]
         self.assertTrue(isinstance(event, FieldUpdatedEvent))
-        self.assertEqual(event.inst, field)
+        self.assertTrue(verifyObject(IFieldUpdatedEvent, event))
+        self.assertEqual(event.object, field)
         self.assertEqual(event.old_value, 0)
         self.assertEqual(event.new_value, 0)
         self.assertEqual(
             [ev.field.__name__ for ev in log],
             ['min_length', 'max_length', 'title', 'description', 'required',
              'readonly'])
+        # BBB, but test this works.
+        self.assertEqual(event.inst, field)
+        marker = object()
+        event.inst = marker
+        self.assertEqual(event.inst, marker)
+        self.assertEqual(event.object, marker)
 
     def test_field_event_update(self):
         from zope.schema import Text
-
+        from zope.interface.verify import verifyObject
         from zope.event import subscribers
+        from zope.schema.interfaces import IFieldUpdatedEvent
         from zope.schema.fieldproperty import FieldUpdatedEvent
         field = Text(
             __name__='testing',
@@ -315,10 +324,17 @@ class FieldPropertyTests(_Base, _Integration):
         self.assertEqual(len(log), 2)
         event = log[1]
         self.assertTrue(isinstance(event, FieldUpdatedEvent))
-        self.assertEqual(event.inst, foo)
+        self.assertTrue(verifyObject(IFieldUpdatedEvent, event))
+        self.assertEqual(event.object, foo)
         self.assertEqual(event.field, field)
         self.assertEqual(event.old_value, u'Bar')
         self.assertEqual(event.new_value, u'Foo')
+        # BBB, but test this works.
+        self.assertEqual(event.inst, foo)
+        marker = object()
+        event.inst = marker
+        self.assertEqual(event.inst, marker)
+        self.assertEqual(event.object, marker)
 
 
 class FieldPropertyStoredThroughFieldTests(_Base, _Integration):
@@ -583,7 +599,7 @@ class FieldPropertyStoredThroughFieldTests(_Base, _Integration):
         self.assertEqual(len(log), 2)
         event = log[1]
         self.assertTrue(isinstance(event, FieldUpdatedEvent))
-        self.assertEqual(event.inst, foo)
+        self.assertEqual(event.object, foo)
         self.assertEqual(event.field, field)
         self.assertEqual(event.old_value, u'Bar')
         self.assertEqual(event.new_value, u'Foo')
@@ -613,7 +629,7 @@ class FieldPropertyStoredThroughFieldTests(_Base, _Integration):
              'readonly'])
         event = log[0]
         self.assertTrue(isinstance(event, FieldUpdatedEvent))
-        self.assertEqual(event.inst, field)
+        self.assertEqual(event.object, field)
         self.assertEqual(event.old_value, 0)
         self.assertEqual(event.new_value, 0)